week-6/requirements.txt
flask
flask-wtf
wtforms
flask-sqlalchemy
mysql-connector-python